Bugs removed , food logger ux improved

+++ app/src/main/java/in/tagbin/mitohealthapp/FoodDetails.java
@@ -269,7 +269,7 @@ public class FoodDetails extends AppCompatActivity {
             long time = date.getTime()/1000L;
             Log.d("timesatmap",""+time);
             setFoodLoggerModel.setTime_consumed(time);
-            setFoodLoggerModel.setAmount(Integer.parseInt(FoodDetailsFrag.quantity));
+            setFoodLoggerModel.setAmount(Float.parseFloat(FoodDetailsFrag.quantity));
             Log.d("model2",JsonUtils.jsonify(setFoodLoggerModel));
             progressBar.setVisibility(View.VISIBLE);
             Controller.updateLogFood(FoodDetails.this,setFoodLoggerModel,data.getId(),mFoodUpdateListener);
@@ -283,7 +283,7 @@ public class FoodDetails extends AppCompatActivity {
              Log.d("timesatmap",""+time);
              setFoodLoggerModel.setTime_consumed(time);
             setFoodLoggerModel.setFlag(1);
-             setFoodLoggerModel.setAmount(Integer.parseInt(FoodDetailsFrag.quantity));
+             setFoodLoggerModel.setAmount(Float.parseFloat(FoodDetailsFrag.quantity));
             Log.d("model",JsonUtils.jsonify(setFoodLoggerModel));
              progressBar.setVisibility(View.VISIBLE);
              Controller.setLogger(FoodDetails.this,setFoodLoggerModel,mFoodLoggerListener);
@@ -298,7 +298,7 @@ public class FoodDetails extends AppCompatActivity {
              setFoodLoggerModel.setMeal_id(data.getMeal_id());
              setFoodLoggerModel.setFlag(1);
             setFoodLoggerModel.setServing_unit(FoodDetailsFrag.servingUnit);
-            setFoodLoggerModel.setAmount(Integer.parseInt(FoodDetailsFrag.quantity));
+            setFoodLoggerModel.setAmount(Float.parseFloat(FoodDetailsFrag.quantity));
             Log.d("model1",JsonUtils.jsonify(setFoodLoggerModel));
             progressBar.setVisibility(View.VISIBLE);
             Controller.setLogger(FoodDetails.this,setFoodLoggerModel,mFoodLoggerListener);
